+++ lib/beyond_api/connection.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'faraday'
+require "faraday"
 
 module BeyondApi
   class Connection
@@ -13,8 +13,8 @@ module BeyondApi
         faraday.options[:timeout] = BeyondApi.configuration.timeout.to_i
         faraday.response :logger, LOGGER, { headers: BeyondApi.configuration.log_headers,
                                             bodies: BeyondApi.configuration.log_bodies }
-        faraday.headers['Accept'] = 'application/json'
-        faraday.headers['Content-Type'] = 'application/json'
+        faraday.headers["Accept"] = "application/json"
+        faraday.headers["Content-Type"] = "application/json"
         faraday.request(:multipart)
         faraday.request(:url_encoded)
         faraday.adapter(:net_http)
@@ -32,8 +32,16 @@ module BeyondApi
         end
         faraday.headers['Accept'] = 'application/json'
         faraday.adapter(:net_http)
-        faraday.basic_auth(BeyondApi.configuration.client_id,
-                           BeyondApi.configuration.client_secret)
+        faraday.request :basic_auth, BeyondApi.configuration.client_id, BeyondApi.configuration.client_secret
+      end
+    end
+
+    def self.multipart
+      Faraday.new(ssl: { verify: true }) do |faraday|
+        faraday.options[:open_timeout] = BeyondApi.configuration.open_timeout.to_i
+        faraday.options[:timeout] = BeyondApi.configuration.timeout.to_i
+        faraday.request :multipart, { flat_encode: true }
+        faraday.adapter Faraday.default_adapter
       end
     end
   end