+++ lib/beyond_api/ext.rb
@@ -5,13 +5,13 @@ class Hash
     result = {}
     each do |key, value|
       result[yield(key)] = case value
-      when Hash
-        value.deep_transform_keys(&block)
-      when Array
-        value.camelize_keys
-      else
-        value
-      end
+                           when Hash
+                             value.deep_transform_keys(&block)
+                           when Array
+                             value.camelize_keys
+                           else
+                             value
+                           end
     end
     result
   end
@@ -31,19 +31,19 @@ class String
   end
 
   def underscore
-    self.gsub(/::/, '/').
-    gsub(/([A-Z]+)([A-Z][a-z])/,'\1_\2').
-    gsub(/([a-z\d])([A-Z])/,'\1_\2').
-    tr("-", "_").
-    downcase
+    gsub(/::/, "/")
+      .gsub(/([A-Z]+)([A-Z][a-z])/, '\1_\2')
+      .gsub(/([a-z\d])([A-Z])/, '\1_\2')
+      .tr("-", "_")
+      .downcase
   end
 
   def camelize(uppercase_first_letter = true)
     string = self
     if uppercase_first_letter
-      string = string.sub(/^[a-z\d]*/) { |match| match.capitalize }
+      string = string.sub(/^[a-z\d]*/, &:capitalize)
     else
-      string = string.sub(/^(?:(?=\b|[A-Z_])|\w)/) { |match| match.downcase }
+      string = string.sub(/^(?:(?=\b|[A-Z_])|\w)/, &:downcase)
     end
     string.gsub(/(?:_|(\/))([a-z\d]*)/) { "#{$1}#{$2.capitalize}" }.gsub("/", "::")
   end
@@ -53,8 +53,8 @@ class Array
   def camelize_keys
     map do |elem|
       case elem
-        when Hash, Array; elem.camelize_keys
-        else; elem
+      when Hash, Array then elem.camelize_keys
+      else; elem
       end
     end
   end