+++ lib/beyond_api/request.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 require "json"
+require "faraday"
+require "beyond_api/utils"
 
 module BeyondApi
   class Request
@@ -9,7 +11,7 @@ module BeyondApi
         define_method(method) do |session, path, params = {}|
           response = BeyondApi::Connection.default.send(method) do |request|
             request.url(session.api_url + path)
-            request.headers['Authorization'] = "Bearer #{ session.access_token }" unless session.access_token.nil?
+            request.headers["Authorization"] = "Bearer #{session.access_token}" unless session.access_token.nil?
             request.params = params.to_h.camelize_keys
           end
 
@@ -21,7 +23,7 @@ module BeyondApi
         define_method(method) do |session, path, body = {}, params = {}|
           response = BeyondApi::Connection.default.send(method) do |request|
             request.url(session.api_url + path)
-            request.headers['Authorization'] = "Bearer #{ session.access_token }" unless session.access_token.nil?
+            request.headers["Authorization"] = "Bearer #{session.access_token}" unless session.access_token.nil?
             request.params = params.to_h.camelize_keys
             request.body = body.camelize_keys.to_json
           end
@@ -34,8 +36,8 @@ module BeyondApi
     def self.upload(session, path, file_binary, content_type, params)
       response = BeyondApi::Connection.default.post do |request|
         request.url(session.api_url + path)
-        request.headers['Authorization'] = "Bearer #{ session.access_token }" unless session.access_token.nil?
-        request.headers['Content-Type'] = content_type
+        request.headers["Authorization"] = "Bearer #{session.access_token}" unless session.access_token.nil?
+        request.headers["Content-Type"] = content_type
         request.params = params.to_h.camelize_keys
         request.body = file_binary
       end
@@ -47,6 +49,21 @@ module BeyondApi
       response = BeyondApi::Connection.token.post do |request|
         request.url(url)
         request.params = params
+      end
+
+      [response.body.blank? ? nil : JSON.parse(response.body), response.status]
+    end
+
+    def self.upload_by_form(session, path, files, params)
+      response = BeyondApi::Connection.multipart.post do |request|
+        request.url(session.api_url + path)
+        request.headers["Authorization"] = "Bearer #{session.access_token}" unless session.access_token.nil?
+        request.options[:params_encoder] = Faraday::FlatParamsEncoder
+        request.params = params.to_h.camelize_keys
+        files = files.split unless files.is_a? Array
+        upload_files = files.map{ |file| Faraday::FilePart.new(File.open(file),
+                                                               BeyondApi::Utils.file_content_type(file)) }
+        request.body = { image: upload_files }
       end
 
       [response.body.blank? ? nil : JSON.parse(response.body), response.status]