+++ lib/beyond_api/resources/token.rb
@@ -13,8 +13,8 @@ module BeyondApi
     def initialize(session)
       @session = session
 
-      raise InvalidSessionError.new("Invalid session") unless session.is_a? BeyondApi::Session
-      raise InvalidSessionError.new("Session api_url cannot be nil") if session.api_url.nil?
+      raise InvalidSessionError, "Invalid session" unless session.is_a? BeyondApi::Session
+      raise InvalidSessionError, "Session api_url cannot be nil" if session.api_url.nil?
     end
 
     def authorization_code(code)
@@ -29,15 +29,16 @@ module BeyondApi
       handle_token_call("client_credentials")
     end
 
-    alias_method :refresh, :refresh_token
-    alias_method :create, :authorization_code
+    alias refresh refresh_token
+    alias create authorization_code
 
     private
 
     def handle_token_call(grant_type, params = {})
+      path = "#{@session.api_url}/oauth/token"
       params.merge!(grant_type: grant_type)
 
-      response, status = BeyondApi::Request.token(@session.api_url + "/oauth/token",
+      response, status = BeyondApi::Request.token(path,
                                                   params)
 
       handle_response(response, status)