+++ spec/beyond_api_spec.rb
@@ -1,9 +1,7 @@
+# frozen_string_literal: true
+
 RSpec.describe BeyondApi do
   it "has a version number" do
     expect(BeyondApi::VERSION).not_to be nil
-  end
-
-  it "does something useful" do
-    expect(false).to eq(true)
   end
 end